Scss mob ver. + navbar orangebackground vid click
