+++ src/sass/main.css
@@ -1,0 +1,1 @@
+body{background-image:url("/water.jpg");background-size:cover;box-sizing:border-box;z-index:-1}.header{position:fixed;right:50px;top:20px;font-family:"Rubik",sans-serif;font-size:1.3rem;font-weight:700}a:link{color:#f80;text-decoration:none}a:hover{border-bottom:solid 3px #f80}a:visited{color:#f80}.buyBtn{font-family:"Rubik",sans-serif;background-color:#3c983c;position:fixed;top:290px;left:40vw;font-size:1.3rem;font-weight:700;padding:10px;color:#f80;width:300px;border-radius:6px;border:none;box-shadow:0 8px 16px 0 rgba(0,0,0,.2),0 6px 20px 0 rgba(0,0,0,.19)}.buyBtn:hover{background-color:#398639}.token{position:fixed;left:120px;top:50px;width:350px}.textBuy{font-family:Georgia,"Times New Roman",Times,serif;color:#f80;font-size:1.2rem;position:fixed;top:360px;left:37vw}.team{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;color:#46ac46;position:fixed;left:200px;top:70px;font-size:2rem}.Dev1pic,.Dev2pic,.Dev3pic,.Dev4pic{border-radius:100px;width:100px;position:fixed;outline:solid 5px #46ac46}.Dev1pic{left:22.5vw;top:200px}.Dev2pic{left:38.5vw;top:200px}.Dev3pic{left:54.5vw;top:200px}.Dev4pic{left:70.5vw;top:200px}.textDev1,.textDev2,.textDev3,.textDev4{font-family:Georgia,"Times New Roman",Times,serif;color:#f80;font-size:1.1rem;font-weight:bold}.textDev1{position:fixed;left:24vw;top:300px}.textDev2{position:fixed;left:40vw;top:300px}.textDev3{position:fixed;left:56vw;top:300px}.textDev4{position:fixed;left:72vw;top:300px}.textDev1a,.textDev2a,.textDev3a,.textDev4a{font-family:Georgia,"Times New Roman",Times,serif;color:#f80;font-size:1rem}.textDev1a{position:fixed;left:23vw;top:330px}.textDev2a{position:fixed;left:39vw;top:330px}.textDev3a{position:fixed;left:55vw;top:330px}.textDev4a{position:fixed;left:71vw;top:330px}.textDev1b,.textDev2b,.textDev3b,.textDev4b{font-family:Georgia,"Times New Roman",Times,serif;color:#f80;font-size:1.2rem}.textDev1b{position:fixed;left:24vw;top:380px}.textDev2b{position:fixed;left:40vw;top:380px}.textDev3b{position:fixed;left:56vw;top:380px}.textDev4b{position:fixed;left:72vw;top:380px}.Roadmap{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;color:#46ac46;position:fixed;left:200px;top:70px;font-size:2rem}.Text{font-family:Georgia,"Times New Roman",Times,serif;color:#f80;font-size:1rem;position:fixed;top:330px;left:33vw}.Text2{font-family:Georgia,"Times New Roman",Times,serif;color:#f80;font-size:1rem;position:fixed;top:330px;left:53vw}.nftDiv{display:flex;align-items:center;flex-wrap:wrap;justify-content:space-evenly;gap:50px;height:50vh;width:80vw;margin-left:auto;margin-right:auto}.Nft{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;color:#46ac46;position:fixed;left:200px;top:70px;font-size:2rem}.Coming1{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;color:orange;position:fixed;left:500px;top:200px;font-size:2.6rem;transform:rotate(-20deg);-webkit-animation:glow;animation:glow;text-shadow:0 0 1px #fff,0 0 2px orange}.nft{background-color:green;height:450px;width:300px;border-radius:10px;display:flex;flex-direction:column}.nft h1{margin-left:auto;margin-right:auto}.nft img{width:200px;height:200px;margin-left:auto;margin-right:auto;border:5px solid #000}.nft:hover{box-shadow:5px 5px #000}.Merch{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;color:#46ac46;position:fixed;left:200px;top:70px;font-size:2rem}.Coming2{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;color:orange;position:fixed;left:35vw;top:150px;font-size:2.6rem;transform:rotate(-20deg);-webkit-animation:glow;animation:glow;text-shadow:0 0 1px #fff,0 0 2px orange;z-index:2}.MerchPic{position:fixed;left:35vw;top:160px;width:450px;z-index:1}.Navbar{height:70px;margin:10px;position:fixed;bottom:15px;left:10px;right:10px;display:flex;flex-direction:row;justify-content:space-evenly;outline:solid 3px #f80;border-radius:10px}.homeBtn,.teamBtn,.roadMapBtn,.nftBtn,.merchBtn{font-family:"Gill Sans","Gill Sans MT","Trebuchet MS",sans-serif;background-color:rgba(0,0,0,0);font-size:1.4rem;font-weight:700;padding:10px;color:#46ac46;width:190px;border:none;z-index:2}.homeBtn:hover{border-bottom:solid 5px #46ac46}.teamBtn:hover{border-bottom:solid 5px #46ac46}.roadMapBtn:hover{border-bottom:solid 5px #46ac46}.nftBtn:hover{border-bottom:solid 5px #46ac46}.merchBtn:hover{border-bottom:solid 5px #46ac46}.NavbarBackground{height:70px;margin:10px;position:fixed;bottom:15px;left:10px;right:10px;display:flex;flex-direction:row;justify-content:space-evenly;border-radius:10px}.backgroundColorHomeBtn,.backgroundColorTeamBtn,.backgroundColorRoadmapBtn,.backgroundColorNftBtn,.backgroundColorMerchBtn{width:190px;background-color:rgba(0,0,0,0);border:none}@media screen and (max-width: 500px){header{right:4px;top:0px}a:link{font-size:.7rem;font-weight:400}#token{left:26px;top:28px;max-width:150px}#buyBtn{top:170px;left:34%;padding:10px;width:150px;font-weight:bold;font-size:.7rem}#textBuy{font-size:.6rem;top:210px;left:24%}#team{left:40px;top:20px;font-size:1.1rem}#dev1pic,#dev2pic,#dev3pic,#dev4pic{width:40px;top:120px;outline:solid 2px #46ac46}#textDev1,#textDev2,#textDev3,#textDev4{top:170px;font-size:.6rem;font-weight:bold}#textDev1a,#textDev2a,#textDev3a,#textDev4a{font-size:.6rem;top:190px}#textDev1b,#textDev2b,#textDev3b,#textDev4b{font-size:.6rem;top:230px}#roadmap{left:40px;top:20px;font-size:1.1rem}#roadmap2{left:24vw;top:120px;width:240px}#text{font-size:.6rem;top:190px;left:23vw}#texte{font-size:.6rem;top:190px;left:53vw}#nft{left:40px;top:20px;font-size:1.1rem}#coming1{left:32%;top:130px;font-size:1.3rem}#merch{left:40px;top:20px;font-size:1.1rem}#coming2{left:30%;top:115px;font-size:1.3rem}#merchPic{position:fixed;left:28%;top:120px;z-index:-1;width:200px}footer{height:60px;margin:10px;bottom:5px;left:5px;right:5px}#homeBtn,#teamBtn,#roadMapBtn,#nftBtn,#merchBtn{font-size:.7rem;font-weight:700}}/*# sourceMappingURL=main.css.map */
